refactor(providers): apply mapping pattern and extract retry decorator
Fixes two PR #129 review comments:

1. Apply mapping dict pattern in __getattr__ for consistency with
   strategies/__init__.py established in PR #127/#128

2. Extract retry decorator to separate module to prevent eager httpx
   loading. BaseLLMProvider now has no httpx dependency - retry moved
   to src/models/providers/retry.py and imported only by lazy-loaded
   provider implementations

- src/models/providers/__init__.py: mapping pattern for all providers
- src/models/providers/base.py: remove retry + httpx imports
- src/models/providers/retry.py: NEW - retry decorator with httpx deps
- src/models/providers/{anthropic,openai,ollama}.py: import retry from new location

Verified: httpx not loaded when importing BaseLLMProvider/ToolCall
All 855 tests pass, mise check clean
