+++ src/models/providers/base.py
@@ -1,18 +1,13 @@
-"""Base LLM provider interface and retry decorator."""
+"""Base LLM provider interface."""
 
-import asyncio
-import functools
 from abc import ABC, abstractmethod
-from collections.abc import AsyncIterator, Awaitable, Callable
-from typing import ParamSpec, TypeVar
+from collections.abc import AsyncIterator
+from typing import TypeVar
 
-import httpx
-from loguru import logger
 from pydantic import BaseModel
 
 from src.metrics.execution import LLMUsageStats
 
-P = ParamSpec("P")
 T = TypeVar("T")
 
 
@@ -36,49 +31,6 @@ class ToolCall(BaseModel):
     id: str
     name: str
     arguments: dict
-
-
-def retry(
-    max_attempts: int = 3,
-    delay: float = 1.0,
-    exceptions: tuple[type[Exception], ...] = (
-        httpx.ConnectError,
-        httpx.TimeoutException,
-        httpx.ReadTimeout,
-        httpx.WriteTimeout,
-        httpx.PoolTimeout,
-        httpx.HTTPStatusError,
-    ),
-) -> Callable[[Callable[P, Awaitable[T]]], Callable[P, Awaitable[T]]]:
-    """Decorator for retrying async functions on transient errors.
-
-    Args:
-        max_attempts: Maximum retry attempts
-        delay: Base delay between retries (exponential backoff)
-        exceptions: Exception types to retry on
-    """
-
-    def decorator(func: Callable[P, Awaitable[T]]) -> Callable[P, Awaitable[T]]:
-        @functools.wraps(func)
-        async def wrapper(*args: P.args, **kwargs: P.kwargs) -> T:
-            last_error: Exception | None = None
-            for attempt in range(max_attempts):
-                try:
-                    return await func(*args, **kwargs)
-                except exceptions as e:
-                    last_error = e
-                    if attempt < max_attempts - 1:
-                        wait_time = delay * (attempt + 1)
-                        logger.warning(f"Retry {attempt + 1}/{max_attempts} after {wait_time}s: {e}")
-                        await asyncio.sleep(wait_time)
-            if last_error is not None:
-                raise last_error
-            msg = "Retry decorator invoked with max_attempts <= 0 or no exceptions captured"
-            raise RuntimeError(msg)
-
-        return wrapper
-
-    return decorator
 
 
 class BaseLLMProvider(ABC):
