+++ src/models/providers/openai.py
@@ -11,7 +11,8 @@ from openai import AsyncOpenAI
 from pydantic import BaseModel, ValidationError
 
 from src.metrics.execution import LLMUsageStats
-from src.models.providers.base import BaseLLMProvider, StructuredOutputError, ToolCall, retry
+from src.models.providers.base import BaseLLMProvider, StructuredOutputError, ToolCall
+from src.models.providers.retry import retry
 
 T = TypeVar("T", bound=BaseModel)
